add be only rolling restart
