+++ app/kubernetes/components/yaml-inspector/yamlInspector.html
@@ -22,6 +22,12 @@
       <pr-icon class="vertical-center" icon="'check'" size="'md'" mode="'success'"></pr-icon> copied
     </span>
 
-    <yaml-replace class="float-right" feature-id="$ctrl.limitedFeature"></yaml-replace>
+    <be-only-button
+      class="float-right"
+      feature-id="$ctrl.limitedFeature"
+      message="'Applies any changes that you make in the YAML editor by calling the Kubernetes API to patch the relevant resources. Any resource removals or unexpected resource additions that you make in the YAML will be ignored. Note that editing is disabled for resources in namespaces marked as system.'"
+      heading="'Apply YAML changes'"
+      button-text="'Apply changes'"
+    ></be-only-button>
   </div>
 </div>
