+++ app/kubernetes/react/views/index.ts
@@ -4,7 +4,7 @@ import { r2a } from '@/react-tools/react2angular';
 import { withCurrentUser } from '@/react-tools/withCurrentUser';
 import { withReactQuery } from '@/react-tools/withReactQuery';
 import { withUIRouter } from '@/react-tools/withUIRouter';
-import { YAMLReplace } from '@/kubernetes/react/views/yamlReplace';
+import { BEOnlyButton } from '@/kubernetes/react/views/beOnlyButton';
 import { IngressesDatatableView } from '@/react/kubernetes/ingresses/IngressDatatable';
 import { CreateIngressView } from '@/react/kubernetes/ingresses/CreateIngressView';
 
@@ -22,8 +22,13 @@ export const viewsModule = angular
     r2a(withUIRouter(withReactQuery(withCurrentUser(CreateIngressView))), [])
   )
   .component(
-    'yamlReplace',
-    r2a(withUIRouter(withReactQuery(withCurrentUser(YAMLReplace))), [
+    'beOnlyButton',
+    r2a(withUIRouter(withReactQuery(withCurrentUser(BEOnlyButton))), [
       'featureId',
+      'heading',
+      'message',
+      'buttonText',
+      'className',
+      'icon',
     ])
   ).name;
